Crate pyo3_macros_backend

This crate contains the implementation of the proc macro attributes

Structs

PyClassArgs

The parsed arguments of the pyclass macro

PyFunctionOptions

PyModuleOptions

WrapPyFunctionArgs

Enums

PyClassMethodsType

The mechanism used to collect #[pymethods] into the type object

Functions

build_derive_from_pyobject

Derive FromPyObject for enums and structs.

build_py_class

build_py_enum

build_py_function

build_py_methods

build_py_proto

get_doc

Collects all #[doc = “…”] attributes into a TokenStream evaluating to a null-terminated string e.g. concat!(“…”, “\n”, “\0”)

process_functions_in_module

Finds and takes care of the #[pyfn(…)] in #[pymodule]

pymodule_impl

Generates the function that is called by the python interpreter to initialize the native module

wrap_pyfunction_impl

wrap_pymodule_impl