Module pyo3_macros_backend::attributes

Modules

kw

Structs

KeywordAttribute

LitStrValue

A helper type which parses the inner type via a literal string e.g. LitStrValue -> parses “some::path” in quotes.

NameLitStr

A helper type which parses a name via a literal string

Functions

get_pyo3_options

is_attribute_ident

take_attributes

Takes attributes from an attribute vector.

take_pyo3_options

Type Definitions

CrateAttribute

For specifying the path to the pyo3 crate.

ExtendsAttribute

FreelistAttribute

FromPyWithAttribute

ModuleAttribute

NameAttribute

TextSignatureAttribute