⚠️ Internal Docs ⚠️ Not Public API 👉 Official Docs Here

Module pyo3_macros_backend::utils

Macros

bail_spanned

Macro inspired by anyhow::bail! to return a compiler error with the given span.

ensure_spanned

Macro inspired by anyhow::ensure! to return a compiler error with the given span if the specified condition is not met.

err_spanned

Macro inspired by anyhow::anyhow! to create a compiler error with the given span.

Structs

DocArgs

PythonDoc

A syntax tree which evaluates to a null-terminated docstring for Python.

Functions

ensure_not_async_fn

get_doc

Collects all #[doc = “…”] attributes into a TokenStream evaluating to a null-terminated string e.g. concat!(“…”, “\n”, “\0”)

get_pyo3_crate

Extract the path to the pyo3 crate, or use the default (::pyo3).

is_python

Check if the given type ty is pyo3::Python.

option_type_argument

If ty is Option, return Some(T), else None.

remove_lifetime

Remove lifetime from reference

replace_self

Replace Self keyword in type with cls

unwrap_group

unwrap_ty_group